feature Problem class
