+++ src/core/Agent.js
@@ -1,4 +1,4 @@
-const tf = require('@tensorflow/tfjs-node');
+//const tf = require('@tensorflow/tfjs-node');
 
 class Agent {
     constructor(name) {
