+++ src/core/AgentController.js
@@ -5,6 +5,7 @@ class AgentController {
         this.world0 = null;
         this.world = null;
         this.actions = [];
+        this.data = {states: [], world : {}}
     }
     /**
      * Setup the configuration for the agent controller
@@ -13,10 +14,10 @@ class AgentController {
     setup(parameter) {
         this.problem = parameter.problem;
         this.world0 = JSON.parse(JSON.stringify(parameter.world));
-        this.data = {world: JSON.parse(JSON.stringify(parameter.world)), states: {}};
+        this.data.world = JSON.parse(JSON.stringify(parameter.world));
         //this.solution = parameter.solution;
         //this.update = parameter.update;
-        this.problemCallback = parameter.callback;
+        //this.callbacks = parameter.callbacks;
         //this.perceptionForAgent = parameter.perceptionForAgent;
     }
     /**
@@ -70,12 +71,12 @@ class AgentController {
         while (!stop) {
             //Creates a thread for every single agent
             Object.values(this.agents).forEach(agent => {
-                if (!this.problem.solution(this.data)) {
+                if (!this.problem.goalTest(this.data)) {
                     agent.receive(this.problem.perceptionForAgent(this.getData(), agent.getID()));
                     let action = agent.send();
                     this.actions.push({agentID: agent.getID(), action});
                     this.problem.update(this.data, action, agent.getID());
-                    if (this.problem.solution(this.data)) {
+                    if (this.problem.goalTest(this.data)) {
                         stop = true;
                     } else {
                         if(this.callbacks.onTurn)
