+++ src/example/Main.js
@@ -1,14 +1,25 @@
 const CleanerProblem = require('./CleanerProblem');
 const CleanerAgent = require('./CleanerAgent');
 
-let myProblem = new CleanerProblem();
+let myProblem = new CleanerProblem({maxIterations: 12});
 
 myProblem.addAgent("Smith", CleanerAgent, {x: 0, y: 2});
 myProblem.solve([
     [0, 0, 0, 0 ],
     [0, 1, 1, -1],
     [0, 1, 0, 0],
-    [0, 0, 0, 1]]);
+    [0, 0, 0, 1]], {
+        onFinish: (result) => {
+            let agentID = result.actions[result.actions.length - 1].agentID;
+            console.log("Winner " + agentID);
+            console.log(result.actions);
+            let world = JSON.parse(JSON.stringify(result.data.world));
+            let agentState = result.data.states[agentID];
+            world[agentState.y][agentState.x] = "X"
+            console.log(world);
+        }, 
+        onTurn0: (result) => { console.log("Turn: " + result) }
+    });
